Move `HttpConfig` to separate file
