+++ src/http/index.ts
@@ -1,5 +1,6 @@
 export * from './dispatch';
 export * from './http.means';
+export * from './http-config';
 export * from './http-error';
 export * from './http-listener';
 export * from './middleware';
